fix stuck-notes bug
Failure to clearTimeout in the body of the first if statement in
throttle() was causing the function calls to happen out of order.

For example, assume the `ms` arg to function is 100ms. If a 'move' call
happens, followed by another 'move' call 90ms later, then a 'silent'
call 20ms after that:

- the silent call will fire immediately since it'd been more than 100ms
  since the initial 'move' call.
- the second 'move' call will fire 80ms after the 'silent' call, since
  the former was scheduled to run 100ms from its invocation attempt.

thus the order of function calls would be "move, silent, move" which
leads to dots that are stuck playing in the other clients.